+++ hydromt/data_adapter/geodataset.py
@@ -1,22 +1,24 @@
 # -*- coding: utf-8 -*-
 from os.path import join
-from .. import gis_utils, io
 import numpy as np
 import pandas as pd
 import xarray as xr
 import geopandas as gpd
 from shapely.geometry import box
+import logging
+from typing import Union, NewType
+from pathlib import Path
 
+from .. import gis_utils, io
 from .data_adapter import DataAdapter
 from ..raster import GEO_MAP_COORD
 
-import logging
 
 logger = logging.getLogger(__name__)
 
-__all__ = [
-    "GeoDatasetAdapter",
-]
+__all__ = ["GeoDatasetAdapter", "GeoDatasetSource"]
+
+GeoDatasetSource = NewType("GeoDatasetSource", Union[str, Path])
 
 
 class GeoDatasetAdapter(DataAdapter):
@@ -92,8 +94,8 @@ class GeoDatasetAdapter(DataAdapter):
         self,
         data_root,
         data_name,
-        bbox,
-        time_tuple,
+        bbox=None,
+        time_tuple=None,
         variables=None,
         driver=None,
         logger=logger,