+++ tests/test_flw.py
@@ -1,5 +1,4 @@
 """Test hydromt.flw submodule"""
-from _pytest.monkeypatch import V
 import pytest
 import numpy as np
 import xarray as xr