+++ docs/changelog.rst
@@ -11,16 +11,31 @@ unreleased
 
 Added
 -----
+- ModelCatalog to discover generic and plugin model classes. `PR #202 <https://github.com/Deltares/hydromt/pull/202>`_
+- Support for 2-dimensional tabular data through the new DataFrameAdapter. `PR #153 <https://github.com/Deltares/hydromt/pull/153>`_
+- API calls to get info about model components and dataset for the dashboard. `PR #118 <https://github.com/Deltares/hydromt/pull/118>`_
+- New submodelclasses in hydromt: ``GridModel``, ``LumpedModel``, ``MeshModel``, ``NetworkModel``
+- Added entrypoints for lumped_model, mesh_model, grid_model
+- New mixin classes created for model specific object: ``GridMixin`` for self.grid, ``LumpedMixin`` for self.response_units, ``MeshMixin`` for self.mesh,
+  ``MapsMixin`` for self.maps
+- New high-level object: self.maps for storing regular rasters data (which can have resolution and / or projection).
+- Maps generic setup methods: ``MapsMixin.setup_maps_from_raster`` and ``MapsMixin.setup_maps_from_rastermapping``
+- Mesh generic setup methods: ``MeshModel.setup_mesh``, ``MeshMixin.setup_maps_from_raster`` and ``MeshMixin.setup_maps_from_rastermapping``
 
 Changed
 -------
+- self.staticgeoms object and methods renamed to self.geoms
+- self.staticmaps object and methods renamed to self.grid and moved into GridModel and GridMixin
 
 Fixed
 -----
+- Bug in backward compatibility of staticgeoms (not read automatically). `Issue #190 <https://github.com/Deltares/hydromt/issues/190>`_
+- Direct import of xarray.core.resample. `Issue #189 <https://github.com/Deltares/hydromt/issues/189>`_
+- Bug in dim0 attribute of raster, removed instead of set to None if no dim0 `Issue #210 <https://github.com/Deltares/hydromt/issues/210>`
 
 Deprecated
 ----------
-
+- self.staticgeoms and self.staticmaps are deprecated.
 
 v0.5.0 (4 August 2022)
 ======================
@@ -30,7 +45,7 @@ Added
 - New raster method for adding gdal_compliant() attributes to xarray object.
 - Function ``to_datetimeindex`` in available preprocess functions for xr.open_dataset in the data adapter.
 - Function ``remove_duplicates`` in available preprocess functions for xr.open_dataset in the data adapter.
-- New ``DataCatalog.from_predefined_catalogs`` and ``DataCatalog.from_archive`` to support predefined data catalogs and archive 
+- New ``DataCatalog.from_predefined_catalogs`` and ``DataCatalog.from_archive`` to support predefined data catalogs and archive
   in a generic way through the data/predefined_catalogs.yml file.
 - Optional formatting for year and month variables in path of data sources.